+++ app/src/main/java/com/dp/cloudcounter/domain/model/Counter.kt
@@ -7,4 +7,4 @@ package com.dp.cloudcounter.domain.model
  * value - it is number of counted item
  * format - for example km or cm or pts, atc.
  */
-data class Counter(val label: String, val value: Float, val format: String)
+data class Counter(val label: String, val value: Int, val format: String)
